refactor(lsm): 移除废弃的MergeIterator实现并优化测试配置
- 移除LSM迭代器中已废弃的MergeIterator相关代码
- 更新测试用例使用GetDefaultOpt获取默认配置
- 统一迭代器选项中的缓存设置参数
- 修复SSTable校验和长度检查逻辑
- 优化跳表迭代器中的字符串处理逻辑
- 更新值日志GC测试中的文件大小限制
- 调整WAL测试文件路径配置方式
- 添加事务中的SetEntry方法实现
- 优化构建器中表大小参数类型处理

+++ lsm/sstBuilder.go
@@ -83,7 +83,7 @@ func newSSTBuilderWithSSTableSize(opt *Options, size int64) *sstBuilder {
 func NewSSTBuilder(opt *Options) *sstBuilder {
 	return &sstBuilder{
 		opt:     opt,
-		sstSize: int64(opt.BaseTableSize),
+		sstSize: opt.BaseTableSize,
 	}
 }
